pyproject.toml
[project]
name = "deepseek-api-tester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.60.2",
    "python-dotenv>=1.0.1",
]
